feat: add filtration to filter
