+++ src/services/databaseDataForFilters.js
@@ -1,0 +1,27 @@
+const utcDateToYMD = require("../helpers/utcDateToYMD");
+
+const databaseDataForFilters = (events) => {
+  const eventCities = [];
+  const eventDates = [];
+  const eventTypes = [];
+
+  events.forEach((event) => {
+    eventCities.push(event.event_address.city);
+    eventDates.push(utcDateToYMD(event.date_time));
+    eventTypes.push(event.event_types[0].event_type);
+  });
+
+  const uniqueEventCities = Array.from(new Set(eventCities));
+  const uniqueEventDates = Array.from(new Set(eventDates));
+  const uniqueEventTypes = Array.from(new Set(eventTypes));
+
+  const result = {
+    eventCities: uniqueEventCities,
+    eventDates: uniqueEventDates,
+    eventTypes: uniqueEventTypes,
+  };
+
+  return result;
+};
+
+module.exports = { databaseDataForFilters };
